Extract persistence to separate project (#136)
