--- aes-core/src/main/java/com/ijioio/aes/core/persistence/PersistenceContext.java
@@ -1,5 +1,0 @@
-package com.ijioio.aes.core.persistence;
-
-public interface PersistenceContext {
-	// Marker interface
-}
