--- aes-core/src/main/java/com/ijioio/aes/core/persistence/PersistenceTransaction.java
@@ -1,10 +1,0 @@
-package com.ijioio.aes.core.persistence;
-
-public interface PersistenceTransaction {
-
-	public void begin() throws PersistenceException;
-
-	public void commit() throws PersistenceException;
-
-	public void rollback() throws PersistenceException;
-}
