--- aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/JdbcPersistenceHandlerProvider.java
@@ -1,10 +1,0 @@
-package com.ijioio.aes.core.persistence.jdbc;
-
-import javax.sql.DataSource;
-
-public interface JdbcPersistenceHandlerProvider {
-
-	public boolean accept(String productName);
-
-	public JdbcPersistenceHandler create(DataSource dataSource);
-}
