--- aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcBooleanPersistenceValueHandler.java
@@ -1,113 +1,0 @@
-package com.ijioio.aes.core.persistence.jdbc.value.handler;
-
-import java.sql.Array;
-import java.sql.JDBCType;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Types;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import com.ijioio.aes.core.TypeReference;
-import com.ijioio.aes.core.persistence.PersistenceException;
-import com.ijioio.aes.core.persistence.jdbc.JdbcPersistenceContext;
-import com.ijioio.aes.core.persistence.jdbc.JdbcPersistenceHandler;
-
-public class JdbcBooleanPersistenceValueHandler extends BaseJdbcPersistenceValueHandler<Boolean> {
-
-	@Override
-	public Class<Boolean> getType() {
-		return Boolean.class;
-	}
-
-	@Override
-	public List<String> getColumns(JdbcPersistenceContext context, JdbcPersistenceHandler handler, String name,
-			TypeReference<Boolean> type, boolean search) {
-		return Collections.singletonList(name);
-	}
-
-	@Override
-	public void write(JdbcPersistenceContext context, JdbcPersistenceHandler handler, TypeReference<Boolean> type,
-			Boolean value, boolean search) throws PersistenceException {
-
-		PreparedStatement statement = context.getStatement();
-
-		try {
-			statement.setObject(context.getNextIndex(), value, Types.BOOLEAN);
-		} catch (SQLException e) {
-			throw new PersistenceException(e);
-		}
-	}
-
-	@Override
-	public void writeCollection(JdbcPersistenceContext context, JdbcPersistenceHandler handler,
-			TypeReference<? extends Collection<Boolean>> type, Collection<Boolean> values, boolean search)
-			throws PersistenceException {
-
-		PreparedStatement statement = context.getStatement();
-
-		try {
-
-			Array array = values != null
-					? statement.getConnection().createArrayOf(JDBCType.valueOf(Types.BOOLEAN).getName(),
-							values.toArray())
-					: null;
-
-			statement.setObject(context.getNextIndex(), array, Types.ARRAY);
-
-		} catch (SQLException e) {
-			throw new PersistenceException(e);
-		}
-	}
-
-	@Override
-	public Boolean read(JdbcPersistenceContext context, JdbcPersistenceHandler handler, TypeReference<Boolean> type,
-			Boolean value) throws PersistenceException {
-
-		ResultSet resultSet = context.getResultSet();
-
-		try {
-			return resultSet.getObject(context.getNextIndex(), Boolean.class);
-		} catch (SQLException e) {
-			throw new PersistenceException(e);
-		}
-	}
-
-	@Override
-	public Collection<Boolean> readCollection(JdbcPersistenceContext context, JdbcPersistenceHandler handler,
-			TypeReference<? extends Collection<Boolean>> type, Collection<Boolean> values) throws PersistenceException {
-
-		ResultSet resultSet = context.getResultSet();
-
-		try {
-
-			Array array = resultSet.getObject(context.getNextIndex(), Array.class);
-
-			if (array != null) {
-
-				Collection<Boolean> collection = getCollection(type, values);
-
-				collection.clear();
-				collection.addAll(Arrays.stream((Object[]) array.getArray()).map(item -> (Boolean) item)
-						.collect(Collectors.toList()));
-
-				return collection;
-
-			} else {
-
-				if (values != null) {
-					values.clear();
-				}
-
-				return null;
-			}
-
-		} catch (SQLException e) {
-			throw new PersistenceException(e);
-		}
-	}
-}
