+++ aes-persistence/src/main/java/com/ijioio/aes/persistence/PersistenceContext.java
@@ -1,0 +1,5 @@
+package com.ijioio.aes.persistence;
+
+public interface PersistenceContext {
+	// Marker interface
+}
