+++ aes-persistence/src/main/java/com/ijioio/aes/persistence/PersistenceTransaction.java
@@ -1,0 +1,10 @@
+package com.ijioio.aes.persistence;
+
+public interface PersistenceTransaction {
+
+	public void begin() throws PersistenceException;
+
+	public void commit() throws PersistenceException;
+
+	public void rollback() throws PersistenceException;
+}
