+++ aes-persistence/src/main/java/com/ijioio/aes/persistence/jdbc/JdbcPersistenceHandlerProvider.java
@@ -1,0 +1,10 @@
+package com.ijioio.aes.persistence.jdbc;
+
+import javax.sql.DataSource;
+
+public interface JdbcPersistenceHandlerProvider {
+
+	public boolean accept(String productName);
+
+	public JdbcPersistenceHandler create(DataSource dataSource);
+}
