+++ aes-persistence/src/main/java/com/ijioio/aes/persistence/jdbc/value/handler/JdbcLongPersistenceValueHandler.java
@@ -1,0 +1,113 @@
+package com.ijioio.aes.persistence.jdbc.value.handler;
+
+import java.sql.Array;
+import java.sql.JDBCType;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Types;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import com.ijioio.aes.core.TypeReference;
+import com.ijioio.aes.persistence.PersistenceException;
+import com.ijioio.aes.persistence.jdbc.JdbcPersistenceContext;
+import com.ijioio.aes.persistence.jdbc.JdbcPersistenceHandler;
+
+public class JdbcLongPersistenceValueHandler extends BaseJdbcPersistenceValueHandler<Long> {
+
+	@Override
+	public Class<Long> getType() {
+		return Long.class;
+	}
+
+	@Override
+	public List<String> getColumns(JdbcPersistenceContext context, JdbcPersistenceHandler handler, String name,
+			TypeReference<Long> type, boolean search) {
+		return Collections.singletonList(name);
+	}
+
+	@Override
+	public void write(JdbcPersistenceContext context, JdbcPersistenceHandler handler, TypeReference<Long> type,
+			Long value, boolean search) throws PersistenceException {
+
+		PreparedStatement statement = context.getStatement();
+
+		try {
+			statement.setObject(context.getNextIndex(), value, Types.BIGINT);
+		} catch (SQLException e) {
+			throw new PersistenceException(e);
+		}
+	}
+
+	@Override
+	public void writeCollection(JdbcPersistenceContext context, JdbcPersistenceHandler handler,
+			TypeReference<? extends Collection<Long>> type, Collection<Long> values, boolean search)
+			throws PersistenceException {
+
+		PreparedStatement statement = context.getStatement();
+
+		try {
+
+			Array array = values != null
+					? statement.getConnection().createArrayOf(JDBCType.valueOf(Types.BIGINT).getName(),
+							values.toArray())
+					: null;
+
+			statement.setObject(context.getNextIndex(), array, Types.ARRAY);
+
+		} catch (SQLException e) {
+			throw new PersistenceException(e);
+		}
+	}
+
+	@Override
+	public Long read(JdbcPersistenceContext context, JdbcPersistenceHandler handler, TypeReference<Long> type,
+			Long value) throws PersistenceException {
+
+		ResultSet resultSet = context.getResultSet();
+
+		try {
+			return resultSet.getObject(context.getNextIndex(), Long.class);
+		} catch (SQLException e) {
+			throw new PersistenceException(e);
+		}
+	}
+
+	@Override
+	public Collection<Long> readCollection(JdbcPersistenceContext context, JdbcPersistenceHandler handler,
+			TypeReference<? extends Collection<Long>> type, Collection<Long> values) throws PersistenceException {
+
+		ResultSet resultSet = context.getResultSet();
+
+		try {
+
+			Array array = resultSet.getObject(context.getNextIndex(), Array.class);
+
+			if (array != null) {
+
+				Collection<Long> collection = getCollection(type, values);
+
+				collection.clear();
+				collection.addAll(Arrays.stream((Object[]) array.getArray()).map(item -> (Long) item)
+						.collect(Collectors.toList()));
+
+				return collection;
+
+			} else {
+
+				if (values != null) {
+					values.clear();
+				}
+
+				return null;
+			}
+
+		} catch (SQLException e) {
+			throw new PersistenceException(e);
+		}
+	}
+}
