+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'advanced-entity-storage'
 include('aes-annotation')
 include('aes-core')
+include('aes-persistence')
 include('aes-persistence-h2')
 include('aes-persistence-postgresql')
 include('aes-processor')
